added some tests, moved validation to another class, added regex in edit texts

+++ app/src/main/java/ru/you11/timeseries/TimeSeries.kt
@@ -8,7 +8,7 @@ import kotlin.collections.HashMap
  */
 class TimeSeries(var name: String,
                  val creationDate: String,
-                 val dataValues: HashMap<String, List<Double>>?,
+                 val dataValues: HashMap<String, List<Float>>?,
                  val xAxisDescription: String? = null,
                  val yAxisDescription: String? = null) {
